refactor: extract web content core package
