+++ packages/web-content-core/package.json
@@ -1,0 +1,26 @@
+{
+  "name": "@openclaw/web-content-core",
+  "version": "0.0.0-private",
+  "private": true,
+  "files": [
+    "dist"
+  ],
+  "type": "module",
+  "main": "./dist/index.mjs",
+  "types": "./dist/index.d.mts",
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.mts",
+      "import": "./dist/index.mjs",
+      "default": "./dist/index.mjs"
+    },
+    "./provider-runtime-shared": {
+      "types": "./dist/provider-runtime-shared.d.mts",
+      "import": "./dist/provider-runtime-shared.mjs",
+      "default": "./dist/provider-runtime-shared.mjs"
+    }
+  },
+  "scripts": {
+    "build": "tsdown src/index.ts src/provider-runtime-shared.ts --no-config --platform node --format esm --dts --out-dir dist --clean"
+  }
+}
