+++ scripts/build-all.mjs
@@ -51,6 +51,7 @@ export const BUILD_ALL_STEPS = [
         "packages/media-understanding-common/package.json",
         "packages/terminal-core/package.json",
         "packages/model-catalog-core/package.json",
+        "packages/web-content-core/package.json",
         "packages/memory-host-sdk/package.json",
         "tsconfig.json",
         "tsconfig.plugin-sdk.dts.json",
@@ -62,6 +63,7 @@ export const BUILD_ALL_STEPS = [
         "packages/media-generation-core/src",
         "packages/media-understanding-common/src",
         "packages/terminal-core/src",
+        "packages/web-content-core/src",
         "src/types",
         "src/video-generation/dashscope-compatible.ts",
         "src/video-generation/types.ts",
