+++ scripts/run-node-watch-paths.mjs
@@ -14,6 +14,7 @@ const RUN_NODE_PACKAGE_SOURCE_ROOTS = [
   "packages/media-generation-core/src",
   "packages/media-understanding-common/src",
   "packages/terminal-core/src",
+  "packages/web-content-core/src",
   "packages/net-policy/src",
 ];
